+++ term_image/cli.py
@@ -27,6 +27,7 @@ from .image.common import _ALPHA_THRESHOLD
 from .logging import Thread, init_log, log, log_exception
 from .logging_multi import Process
 from .tui.widgets import Image
+from .utils import OS_IS_UNIX
 
 
 def check_dir(
@@ -1089,9 +1090,7 @@ NOTES:
     )
     opener.start()
 
-    os_is_unix = sys.platform not in {"win32", "cygwin"}
-
-    if os_is_unix and not args.cli:
+    if OS_IS_UNIX and not args.cli:
         dir_queue = mp_Queue() if logging.MULTI and args.checkers > 1 else Queue()
         dir_queue.sources_finished = False
         check_manager = Thread(
@@ -1116,7 +1115,7 @@ NOTES:
             if args.cli:
                 log(f"Skipping directory {source!r}", logger, verbose=True)
                 continue
-            if not os_is_unix:
+            if not OS_IS_UNIX:
                 dir_images = True
                 continue
             dir_queue.put(("", [], source, 0))
@@ -1127,7 +1126,7 @@ NOTES:
     for _ in range(args.getters):
         url_queue.put(None)
     file_queue.put(None)
-    if os_is_unix and not args.cli:
+    if OS_IS_UNIX and not args.cli:
         if logging.MULTI and args.checkers > 1:
             dir_queue.sources_finished = True
         else:
@@ -1136,14 +1135,14 @@ NOTES:
     for getter in getters:
         getter.join()
     opener.join()
-    if os_is_unix and not args.cli:
+    if OS_IS_UNIX and not args.cli:
         check_manager.join()
 
     notify.stop_loading()
     while notify.is_loading():
         pass
 
-    if not os_is_unix and dir_images:
+    if not OS_IS_UNIX and dir_images:
         log(
             "Directory sources skipped, not supported on Windows!",
             logger,
@@ -1226,7 +1225,7 @@ NOTES:
             # or padding width/height checks.
             except ValueError as e:
                 notify.notify(str(e), level=notify.ERROR)
-    elif os_is_unix:
+    elif OS_IS_UNIX:
         notify.end_loading()
         tui.init(args, images, contents)
     else: