+++ term_image/tui/__init__.py
@@ -12,6 +12,7 @@ import urwid
 
 from .. import logging
 from ..config import max_notifications
+from ..utils import lock_input
 from . import main
 from .main import process_input, scan_dir_grid, scan_dir_menu, sort_key_lexi
 from .render import image_render_queue, manage_grid_renders, manage_image_renders
@@ -67,6 +68,9 @@ def init(
         daemon=True,
     )
 
+    urwid.raw_display.Screen.get_available_raw_input = lock_input(
+        urwid.raw_display.Screen.get_available_raw_input
+    )
     main.loop.screen.clear()
     main.loop.screen.set_terminal_properties(2**24)